requirements.txt
fastapi
uvicorn[standard]
openai-agents
python-dotenv
